rteval: kcompile: Fix for no numactl and no cpulist
Constructing a command to compile the kernel as a load can fail
in certain cases such as when numactl is not available and the user
doesn't pass a cpulist.

Fix this for both the runcmd and the cleancmd by not assuming that
numactl is available or that taskset will be used.

Reported-by: Punit Agrawal <punit1.agrawal@toshiba.co.jp>
Signed-off-by: John Kacur <jkacur@redhat.com>