rteval: services.py: Fix incorrect detection of container environment
rteval mistakenly detects that it is running inside a container even
though it is running directly on the host. On further investigation
this was found to be due to change in behaviour around byte strings
and strings when going from python2 to python3.

In python3 process output (stdout, stderr) for processes executed via
Popen() is retuned as byte strings and byte strings are not equivalent
to strings, i.e., b'' == '' is False.

Update all call sites for Popen() to include "text=True"
argument - this ensures that the process output is returned as a
string.

Signed-off-by: Punit Agrawal <punit1.agrawal@toshiba.co.jp>
Signed-off-by: John Kacur <jkacur@redhat.com>