rteval: Don't run hackbench by default on low memory systems
commit cb8263770e4f5834a43db6be8ffb55ffd7f876c9
Changed rteval to allow running on low memory systems by default because
on some machines, such as certain Arm boards, it isn't a problem.

However, on our automated testing systems it is causing problems.

This patch disables running hackbench by default on a low memory system.
In order to allow people to run on a lowmem system where this isn't a
problem, this patch also creates the option
--hackbench-runlowmem=True|False

A user who wants hackbench to run in rteval on a low memory system, only
needs to specify

--hackbench-runlowmem=True

Signed-off-by: John Kacur <jkacur@redhat.com>