rteval: kcompile: Fix regular expression to match kernel prefix
If the user specifies a kernel to compile as a load other than the
default kernel, the kernel prefix is obtained with a regular expression.

Currently the regular expression does not accomodate two digit numbers
in the kernel version.
Fix that regular expression to accomodate different kernel versions,
with room to grow for the future.

Signed-off-by: John Kacur <jkacur@redhat.com>