rteval: fix regex in kcompile to calculate patch number for kernel prefix
Updated the regular expression to find the kernel prefix of the source
file. If there are kernel source files with the same major.minor version
but different patch numbers, kcompile would pick up the first one that
is listed based on the kernel prefix and ignore the patch number.
Example: linux-5.17.11.tar.xz and linux-5.17.13.tar.xz. If 5.17.13 had
to be tested, kcompile could pick up linux-5.17.11.tar.xz

Signed-off-by: Manasi Godse <magodse@redhat.com>
Signed-off-by: John Kacur <jkacur@redhat.com>