rteval: default_config_search: Return None on failure
If False is used without checking, it will be interpreted as stdin,
hanging rteval waiting for input.  OTOH, None will cause os.path.exists()
to throw an exception, so we need to check both the name and the existence
separately anyway.  However, this is a better failure mode than hanging
on stdin if the user of the filename fails to check both.

Signed-off-by: Crystal Wood <crwood@redhat.com>
Signed-off-by: John Kacur <jkacur@redhat.com>