rteval: Remove 30 second "settling" period
Waiting for 30 seconds for loads to "settle down" adds unneccesary delay
to very short runs, besides being a giant hack.  Load modules already
have the opportunity to do setup beforehand.

Even if there are a few seconds before the loads get to their expected
"heaviness", that shouldn't meaningfully affect the output except for:
 - very short runs where you either
   - don't care about the latency because you're debugging rteval itself, or
   - are debugging a latency that reliably shows quickly, in which case
     consistent behavior is enough
 - latency spikes that only show up during load startup activity, in
   which case measuring it is a good thing
 - minimum latency values, which are not really the point of rteval.

...and the 30 second delay is *especially* useless if loads are disabled.
Currently there's no official way to do that (as far as I can find), but
it can be done by disabling all but stressng in the conf file, and not
supplying stressng options on the command line.

Signed-off-by: Crystal Wood <crwood@redhat.com>
Signed-off-by: John Kacur <jkacur@redhat.com>