rteval: Add summary reporting for timerlat
This adds an rteval section to the xsl file and generates a timerlat
report at the end of a run. To use it edit your rteval.conf file
to comment out cyclictest and uncomment timerlat

Another interesting thing you can do is uncomment both of them and
get a report from both in one run.

You can also use this with the summary report, for example

rteval -Z rteval-20240426-3/summary.xml

will work with timerlat now. (or both cyclictest and timerlat at the
same time.)

Signed-off-by: John Kacur <jkacur@redhat.com>