rteval: Fix default measurement config
Commit 761741d15d08e ("rteval: rteval-cmd: Some style changes suggested by pylint-3")
accidentally indented the check for missing measurement config, such that
it became part of the missing load conditional.  This was harmless if
the cnofig file was missing entirely, but not if the config file
exists but the measurement section is empty.

Also, remove timerlat from the default as it conflicts with cyclictest,
and we're not quite ready to make timerlat the default.

Signed-off-by: Crystal Wood <crwood@redhat.com>